Basic Player Movement
Scripted player movement and jump. Added Foreground and Player layers.

Specifically for the jump, the game checks if the player is touching the ground or not (on top of whether or not the space bar is being pressed).

Furthermore, this commit adds a temporary ground block that I might make into a prefab later.